ENH: form for creating new products (#3)
Closes #2 

Adds a form for creating new products, which is opened and filled when scanning products with the barcode scan. Changed navigation to provide two different options for adding products- scan or manual. This should have a clearer UX in a future commit to show that they are both resulting in a product being added.

Includes unit tests which provide a mock of the API response from camera scanning and test that the form is filled correctly, as well as an error message appearing if the product was not found in the API. Removed unused files.